requirements.txt
absl-py==0.8.1
astor==0.8.0
audioread==2.1.8
cachetools==3.1.1
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
cycler==0.10.0
decorator==4.4.1
gast==0.2.2
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
idna==2.8
joblib==0.14.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
librosa==0.7.1
llvmlite==0.30.0
Markdown==3.1.1
matplotlib==3.1.1
numba==0.46.0
numpy==1.17.4
oauthlib==3.1.0
opt-einsum==3.1.0
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
pyparsing==2.4.5
python-dateutil==2.8.1
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.0
scikit-learn==0.21.3
scipy==1.3.2
six==1.13.0
SoundFile==0.10.2
tensorboard==2.0.1
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
tqdm==4.38.0
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
